requirements.txt
dash==2.14.2
opencv-python==4.7.0.72
tensorflow==2.17.1
torch==2.0.1
torchvision==0.15.2
ultralytics==8.0.196
numpy==1.24.3
Pillow==9.5.0
pandas==2.0.3